RA-1258 moved validation test files to fit consistent structure.
